Move util functions to where they are used
